:recycle: move internal module
